+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 3.7.1 under development
 
-- no changes in this release.
+- Enh #220: Add `non-empty-string` psalm type of `Html::generateId()` method result (@vjik)
+- Enh #220: Add `non-empty-string|null` psalm type of `Tag::id()` method parameter (@vjik)
 
 ## 3.7.0 September 18, 2024